Big sur bugfix

+++ ngu/hash.c
@@ -274,10 +274,10 @@ void ripemd160(const uint8_t *msg, int msglen, uint8_t digest[20])
     }
 
 #if !defined(MP_ENDIANNESS_LITTLE)
-#error "untested; suspect endian challenge here"
+#error "untested; suspect endian challenges here"
 #endif
 
-    if(((uint32_t)digest) & 0x3) {
+    if(((size_t)digest) & 0x3) {
         // unaligned case
         uint32_t    ctx[5];
